Remove constructor logic, change append
I had previously started instances of LinkedList with a value property === null and a next property === null, I have concluded that this will not work and we instead need to start with an empty object and have checks for it inside the methods.
With those changes in effect i have changed the append method to check if the object is empty and if it is then to add the properties of the Node instance into this LinkedList object instead, otherwise if there is a pre existing node to then add on this new node to the end of the linked list.