Add changes to toString, change type to module
Changing type to module in package json file allows node to understand that these are es6 modules and results in it processing them correctly.
The changes in toString in the LinkedList class results in a working version which uses a while loop to build template strings, ands them into an array, and then joins them into a string.