Add head property to LinkedList class
This property will point to the first node in the linked list, prior to this I was using an empty object or the first node to populate the linked list, which now makes no sense what I was doing. Sadly I asked ai a different question and it suggested to use the head property and it made complete sense right away.
Next I will change my current methods to use this head property of the linked list instances.