Add logic for at method
This method given an index position will return null if a node does not exist at that position or returns the node. The code differs from previous methods because we have to return the node, so in the loop we build onto a temp variable by referencing each next property accessed, and returning this after the iteration allows us to return the specific node.