Add logic for insertAt method
Added logic for insertAt method to work, the way I have it right now the method will console.log and error message if the index value they gave is too far from the current nodes in the list, and I'm wondering if it would be better if I just called append instead and add their node to the back. I guess not because they can call append and I don't want to force them into something they didn't specifically ask for.