Add find method logic, change edge cases, etc
insertAt and removeAt did not have logic for the ability to remove the first node, I've gone ahead and added that via else if statements.
toString's while loop used a stopping condition of when the current nodes next property === null, which would also stop the loop from adding the current nodes data into the array of strings. By changing the stopping condition to when the currentNode was not equal to null allows our function to add node values to the string array up to the last node.
I also added logic for the find method.
I'll go through all of these methods and reevaluate them to make sure they are working how I planned for them to work before pushing.
Change the pop method because it also had problems with edge cases, instead of using an if else to check what properties the last node had, i've added an else if to check when the only node is being popped off.
Changed find to use a value and not an index, which resulted in rewriting most of the logic. It was actually wrongly written in addition to using the wrong metric, the while loop wasn't incrementing the indexCount nor advancing the currentNode to be the next node, I'm perplexed how it wasn't creating an endless loop but ok I've fixed it.
I'm having a problem with the head method not working and am wondering if I should either change its name or change the head property's name.