Change the head method name to getHead
I could have either changed the property name of head in the class but I didn't want to have to replace all instances of that property in all of our predefined methods, so I've gone ahead and made this method name change. The method now works. I previously thought that javascript would know the difference between a property that wasn't in reference to a method and a method property, so I've learned something new today.