Change insertAt method to correctly append
How this method was written did not catch correctly when a user wanted to append to a list a new node. Even if we have append, I'd still like to give them the ability to do the action.
And I guess I could have just called append at that point but right now with this assignment it'd feel like cheating.